new changes 19102013

+++ src/com/jakubrojcek/BeliefQ.java
@@ -34,7 +34,7 @@ public class BeliefQ {
         this.q = q;
     }
 
-    public void setDiff(float diff) {
+    public void setDiff(double diff) {
         this.diff = diff;
     }
